# Updated RandomState (deprecated from numpy) to default_rng (Generator)