--- a/gensim/models/word2vec.py
+++ b/gensim/models/word2vec.py
@@ -384,7 +384,7 @@ class Word2Vec(utils.SaveLoad):
 
         self.window = int(window)
         self.shrink_windows = bool(shrink_windows)
-        self.random = np.random.default_rng(seed)
+        self.random = np.random.RandomState(seed)
 
         self.hs = int(hs)
         self.negative = int(negative)