--- a/gensim/test/test_ldamodel.py
+++ b/gensim/test/test_ldamodel.py
@@ -31,9 +31,9 @@ corpus = [dictionary.doc2bow(text) for text in common_texts]
 
 
 def test_random_state():
-    testcases = [np.random.seed(0), None, np.random.RandomState(0), 0]
+    testcases = [np.random.seed(0), None, np.random.default_rng(0), 0]
     for testcase in testcases:
-        assert(isinstance(utils.get_random_state(testcase), np.random.RandomState))
+        assert(isinstance(utils.get_random_state(testcase), np.random.Generator))
 
 
 class TestLdaModel(unittest.TestCase, basetmtests.TestBaseTopicModel):
@@ -51,8 +51,8 @@ class TestLdaModel(unittest.TestCase, basetmtests.TestBaseTopicModel):
         assert_allclose(self.model.get_topics(), model2.get_topics(), rtol=1e-5)
 
         # properly continues training on the new state
-        self.model.random_state = np.random.RandomState(0)
-        model2.random_state = np.random.RandomState(0)
+        self.model.random_state = np.random.default_rng(0)
+        model2.random_state = np.random.default_rng(0)
         self.model.passes = 1
         model2.passes = 1
         self.model.update(self.corpus)