@SagarDollin 2021-08-31
Things achieved in this PR:

1. I have resolved all the dependencies on RandomState.
2. The code is still backward compatible in the sense, we can load a pre-trained model that relies on RandomState instead of the Generator and still be able to run it.