@piskvorky 2022-02-19
Thanks for investigating! TBH I'm not a fan of all the if-then logic. It will be really hard to maintain.

When does numpy actually drop the deprecated `RandomState`? We should make a hard cliff and switch over, without the ids.